finished project

+++ js/stories.js
@@ -20,13 +20,13 @@ async function getAndShowStoriesOnStart() {
  */
 
 function generateStoryMarkup(story) {
-  // console.debug("generateStoryMarkup", story);
+  console.debug("generateStoryMarkup", story);
 
   const hostName = story.getHostName();
   if (currentUser && currentUser.favorites.some(val => val.storyId === story.storyId)) { // returns true is it's previously favorited
-    console.log(currentUser.favorites, 'came out true');
     return $(`
     <li id="${story.storyId}">
+      <hr>
       <a href="${story.url}" target="a_blank" class="story-link">
         ${story.title}
       </a>
@@ -34,14 +34,13 @@ function generateStoryMarkup(story) {
       <small class="story-author">by ${story.author}</small>
       <small class="story-user">posted by ${story.username}</small>
       <i class="fa-star fa-solid"></i>
-      <i class="fa-solid fa-trash-can trash"></i>
-      <hr>
     </li>`);
 
   }
   else { // not favorited
     return $(`
       <li id="${story.storyId}">
+        <hr>
         <a href="${story.url}" target="a_blank" class="story-link">
           ${story.title}
         </a>
@@ -49,8 +48,6 @@ function generateStoryMarkup(story) {
         <small class="story-author">by ${story.author}</small>
         <small class="story-user">posted by ${story.username}</small>
         <i class="fa-regular fa-star"></i>
-        <i class="fa-solid fa-trash-can trash"></i>
-        <hr>
       </li>`);
   }
 }
@@ -66,7 +63,15 @@ function putStoriesOnPage() {
   for (let story of storyList.stories) {
     const $story = generateStoryMarkup(story);
     $allStoriesList.append($story);
+
+    // check if user is logged in and if story was created by user. Then add optional trash box
+    if (currentUser && story.username == currentUser.username) {
+      let trash = document.createElement('i');
+      trash.className = 'fa-solid fa-trash-can trash';
+      document.getElementById(`${story.storyId}`).appendChild(trash)
+    }
   }
+
 
   $allStoriesList.show();
   
@@ -85,22 +90,16 @@ function putStoriesOnPage() {
 async function newStory(evt) {
   console.debug("new story", evt);
   evt.preventDefault();
-  console.log(currentUser);
   if (!currentUser) { // Doesn't let user post story if they're not logged in
     alert('Please login before adding a new story.')
   }
   // grab the new story info and make new Story instance
   let newStoryInput = await storyList.addStory(currentUser, {title: $('#story-title').val(), author: $('#story-author').val(), url: $('#story-url').val()});
-  console.log('newstory instance', newStoryInput);
-  // console.log(`${newStoryInput instanceof Story}`);
-  // $allStoriesList.append(generateStoryMarkup(newStoryInput));
 
   hidePageComponents();
   
   getAndShowStoriesOnStart();
   
-  console.log('newstory instance', newStoryInput);
-
 }
 
 $newStoryForm.on("submit", newStory);
@@ -108,28 +107,27 @@ $newStoryForm.on("submit", newStory);
 
 // Remove story from user click
 async function removeStory(evt) {
-  console.log('removing story');
-  //await storyList.deleteStory(currentUser, evt.target.parentElement.id);
+  await storyList.deleteStory(currentUser, evt.target.parentElement.id);
   evt.target.parentElement.remove();
-  //console.log(currentUser, storyList);
-
 }
 
 // add/remove story to favorites
 async function toggleFavs(evt) {
-  evt.target.classList.toggle('fa-solid'); // class of favorite
-  evt.target.classList.toggle('fa-regular'); // class of non-favorite
-  let storyID = evt.target.parentElement.id;
-  if (evt.target.classList.value.indexOf('fa-solid') !== -1 ) { // add to favorites
-    console.log('added', currentUser, storyID);
-    let user = await User.addToFavorites(currentUser, storyID);
-    currentUser.favorites.push({storyId: storyID});
-  } else { // removes from favorites
-    console.log('removed')
-    let user = await User.deleteFromFavorites(currentUser, storyID);
-    currentUser.favorites.pop();
-    console.log(currentUser.favorites);
+  if (currentUser) {
+    evt.target.classList.toggle('fa-solid'); // class of favorite
+    evt.target.classList.toggle('fa-regular'); // class of non-favorite
+    let storyID = evt.target.parentElement.id;
+    if (evt.target.classList.value.indexOf('fa-solid') !== -1 ) { // add to favorites
+      let user = await User.addToFavorites(currentUser, storyID);
+      currentUser.favorites.push({storyId: storyID});
+    } else { // removes from favorites
+      let user = await User.deleteFromFavorites(currentUser, storyID);
+      currentUser.favorites = currentUser.favorites.filter(st => st.storyId !== storyID);
+    }
+  } else {
+    alert("Please login first.")
   }
+
 }
 
 // display page of favs when the nav button is clicked
@@ -137,10 +135,9 @@ function displayFavs () {
   if (!currentUser) {
     alert('Please login first.')
   } else {
-    console.log('displaying favs now')
     hidePageComponents();
     putStoriesOnPage()
-    $('.fa-regular.fa-star').parent().remove(); // don't display non-favorite stories
+    $('.fa-regular.fa-star').parent().remove(); // remove non-favorite stories
   }
 }
